refactor: extract SystemLaunchAtLoginService.swift from LaunchAtLoginService.swift
Splits the `SystemLaunchAtLoginService` (the SMAppService-backed
implementation of LaunchAtLoginControlling) out of
LaunchAtLoginService.swift into its own file, mirroring #737
TestingLaunchAtLoginService. LaunchAtLoginService.swift retains the
LaunchAtLoginStatus enum + LaunchAtLoginControlling protocol.

Byte-identical move of lines 63-96. Both files import Foundation +
ServiceManagement.

LaunchAtLoginService.swift: 112 → 76 lines. Tests: 667.
