+++ Sources/BugNarrator/Services/SystemLaunchAtLoginService.swift
@@ -1,0 +1,37 @@
+import Foundation
+import ServiceManagement
+
+struct SystemLaunchAtLoginService: LaunchAtLoginControlling {
+    func currentStatus() -> LaunchAtLoginStatus {
+        guard #available(macOS 13.0, *) else {
+            return .unavailable
+        }
+
+        switch SMAppService.mainApp.status {
+        case .notRegistered:
+            return .disabled
+        case .enabled:
+            return .enabled
+        case .requiresApproval:
+            return .requiresApproval
+        case .notFound:
+            return .notFound
+        @unknown default:
+            return .unavailable
+        }
+    }
+
+    func setEnabled(_ enabled: Bool) throws -> LaunchAtLoginStatus {
+        guard #available(macOS 13.0, *) else {
+            return .unavailable
+        }
+
+        if enabled {
+            try SMAppService.mainApp.register()
+        } else {
+            try SMAppService.mainApp.unregister()
+        }
+
+        return currentStatus()
+    }
+}
